corrected final changes - setted up the bootJar command for Gradle in the build.gradle file and updated the process to install .jar files. This allows launching analogous applications thro PowerShell for Windows for the same both profiles and also separates the .jar files for convenient execution. The executable file is now split, enabling the use of one or two profiles (optional) for standalone launches via the local machine's OS.
