+++ src/main/resources/application-additional.properties
@@ -1,0 +1,1 @@
+server.port=8092
