+++ src/main/resources/application.properties
@@ -8,6 +8,5 @@ spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 
 path.to.avatars.folder=avatars
-server.port=8090
-spring.profiles.active=dev
 #default port number
+server.port=8090
